⇞ bump crate versions
We are going to release a new version of Viceroy.

At some point we will document our release process (#25), versioning
policy (#26), and so on. For now, we'll do this by the seat of our pants
a bit.

We already have a 0.2.0 release, so the first order of business is going
to be bumping the versions.