Have CI use --locked flag for cargo commands
We were already making sure that our releases checked that Viceroy was
using an updated Cargo.lock file, but it would be better if we checked
this on every PR as well so that we can release code without any
hiccups. This commit changes our Makefile which is used as part of CI to
use the --locked flag for the cargo commands that could use it so that
each PR is checked for an up to date lockfile.

Closes #15