Add a Code of Conduct to Viceroy
While this commit could really just be the title I think it's important
to expand on the "why have a Code of Conduct?" question that gets asked.
Sometimes in good faith, but unfortunately often in bad faith. For
those who think a Code of Conduct is bad and try to open up issues on
the issue tracker about this addition we'll kindly ask you to leave and
lock them. For those who are legitimately curious do read on as to the
why.

Often times tech is a pretty hostile space for many people, especially
with Open Source Software (OSS), but also at companies themselves. Now
you might be thinking "Well I've never experienced harrassment or rude
people in tech" and that's great. Honestly I'd hope no one would be
subject to that kind of treatment. That being said it's annecdotal and
if you spend time talking to minorities in tech (whether it's sex, race,
gender identity or something else) you'll find this is often not the
case. Personally I've barely experienced anything at the level of
harrasment that coworkers or friends in the industry have experienced.
We want a Code of Conduct so that we have a mechanism to enforce safety
within our community. While we'd like to be open for everyone we don't
want to be open at the expense of the well being of others. Working on
code just isn't fun if all you're going to recieve is angry threats from
developers and worst of all if we tolerate that kind of behavior people
will leave and they won't make a fuss of it they'll just go.

Worse, people see this behavior being tolerated and won't want to be
part of that community or participate in it. By tolerating intolerance
by others, a community either diminishes or becomes intolerant itself.
This is where the paradox of intolerance comes in, in order to have a
tolerant and welcoming community we must be intolerant of intolerance.
This is what the Code of Conduct is for! We don't want people who are
rude or make people feel unwelcome. We want others who are kind, caring,
and make others want to be here and feel supported. A Code of Conduct is
by no means perfect and only one part of this, but it's the bare minimum
needed to have a healthy community.

If you feel excluded because of the Code of Conduct, I'd ask you to
think why. Is it because you can't be rude towards someone you probably
don't know online? Is it because in a public professional space you
can't swear? Is it so bad that you can't do these things? The Code of
Conduct doesn't ask for a lot and while you might say "We're all adults
here, we don't need one", consistently that just hasn't been the case in
this industry. We're always open to contributions we just want to have a
certain level of professionalism.

At Fastly we also have our own values and code of conduct and part of
this includes the values "We are trustworthy" and "We are good people"
and if we want others to trust us and if we want to be good people
looking out for our community's well being over the hurt feelings of an
individual who feels excluded because they can't be rude is a tradeoff
we'll make in a heartbeat.

In summary while a Code of Conduct does exclude people, it excludes the
kinds of people we'd rather not interact with and instead invites the
people who we would want to interact with into the community. We want
people who care and who are kind and compassionate. If you've read this
long do come join us. A better world of software where we are nice to
each other, where we build each other up, and build some really neat
things is possible.

[fastly-coc]: https://www.fastly.com/code-of-business-conduct-and-ethics/