Add local dictionary support
This reads a new section from the `local_server` part of fastly.toml named `dictionaries`.

The `local_server.dictionaries` section is an array of tables with two fields: name, and file.

The name field is a string which is used as the name of the dictionary and is used via the Dictionary.open method. This field uses the same validation rules asthe Fastly API (Name must start with alphabetical and contain only alphanumeric, underscore, and whitespace).

The file field is a string which is a file-path pointing to a JSON file which contains all the dictionary items within an object.

An example of a full local_server section of fastly.toml now looks like this:
```toml
[local_server]
              [local_server.backends]
                [local_server.backends.dog]
                url = "http://localhost:7878/dog-mocks"
              [local_server.dictionaries]
                    [local_server.dictionaries.a]
                    name="a"
                    file="./a.json"
```