Add closing functionality to Viceroy for Handles
This adds the ability for Viceroy to handle closing handles for
Response, Request, Body, and StreamingBody Handle types via the close
host calls that were added to the public ABI and internally for C@E
which has rolled out to the fleet. While this does not include any SDK
updates for the tests, I've tested this out with the new version of the
Rust SDK to make sure things work out correctly which helped find a bug
or with how I implemented it with Viceroy. With this we can still work
with old versions of the SDK and handle the new SDK as well.

Closes #64