Make URI and Host header semantics more explicit
Previously, we were relying on subtleties of Hyper's behavior to determine the final URI and Host header for outbound requests. Unfortunately, that behavior is sensitive to a number of factors, e.g. whether we happen to be sending a request via an HTTP2 connection.

This commit changes Viceroy to canonicalize the request URI and Host header before it ever reaches Hyper. In most cases this behavior will match what Hyper would do anyway.

Fixes #89