✨ tests: install a pretty tracing subscriber
this adds a few lines to `viceroy_cli::tests::common::Test::against`,
which is used to run test fixtures.

this initializes and installs a tracing subscriber for use in tests,
configured to record any events (i.e., filtering at the `trace` level)
that occur in `viceroy-lib`.

because the `with_test_writer` method is used, the subscriber will use a
writer that is compatible with `cargo test`'s input capturing, and these
verbose logs will not be seen if a test finishes successfully.

an example snippet of this output is shown below:

```
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
test example_test_case ... FAILED

failures:

---- example_test_case stdout ----
  Dec 16 10:39:25.802  INFO viceroy_lib::execute: handling request GET http://127.0.0.1/
    at lib/src/execute.rs:252
    in viceroy_lib::execute::request with id: 0

  Dec 16 10:39:25.804 TRACE viceroy_lib::wiggle_abi::fastly_abi: abi_version: 1
    at lib/src/wiggle_abi.rs:59
    in viceroy_lib::wiggle_abi::fastly_abi::wiggle abi with module: "fastly_abi", function: "init"
    in viceroy_lib::execute::request with id: 0

  Dec 16 10:39:25.804 TRACE viceroy_lib::wiggle_abi::fastly_abi: result: Ok(())
    at lib/src/wiggle_abi.rs:59
    in viceroy_lib::wiggle_abi::fastly_abi::wiggle abi with module: "fastly_abi", function: "init"
    in viceroy_lib::execute::request with id: 0
```