Update ABI and stub out new function
This commit updates the ABI of C@E to the latest version so that v0.8.1
Rust SDK users will not have any crashes due to a function that does not
exist. While this hostcall exists on the fleet and had been updated
internally, we hadn't updated Viceroy to have access to it as well
through a public update. This commit brings Viceroy in line with the
current Rust SDK release and stubs out the
`auto_decompress_response_set` function for `Request` that can be
implemented for local testing in a commit. For now this will stop
clients from crashing when testing locally.

Fixes #110