Add a test case for automatically decompressing response data.
These tests are currently broken, but should succeed once decompression
works. They should also catch a couple error cases, although it's
unclear if the behavior I'm testing for in the case that someone sends
a bad response (one marked as compressed that actually isn't) is ideal.
These will probably need to be extended to support streaming / async
cases in the future.