Clean up the decompression polling logic, using BytesMut instead of Vec.
This appears to be a bit more standardized, and provides a nicer way to
carve off early chunks, as well. Based on testing, it's possible that
this will consistently generate an empty chunk at the end of the
decompression run, but that should be managed correctly by other parts
of the system.