Wire everything up, and get the basic tests working.
This is relatively straightforward, but more thought should go into the
error case in which someone marks data as gzip-compressed but it isn't.
The system remains whole after this, but the observable behavior is odd
in that it gets marked as StatusCode::OK but taking the body leads to
uncatchable failures (I think).