Put the whole `CompressedHttpBody` back into the queue, not just the `hyper::Body`.
A bit of an oops on the previous implementations. This requires slightly
more careful handling of lifetimes to avoid copying. It also only puts
the current subchunk back into the queue in the non-error cases, as
opposed to previous versions which would put error cases back up to get
more errors.