Switch to a more open-ended `ViceroyRequestMetadata` for holding extension data.
Other folks should be able to just pile on to this structure to add
additional information in the future, without much pain, and without
having to add a new type for every item. There's a benign clippy warning
in `req_impl.rs` that'll be nicely resolved when someone does so, too.