𝟎 only log non-zero exit status codes
if a program invokes e.g., `std::process:exit(0)`, this should not cause
confusing diagnostic messages as reported in #127:

```
  Jan 1 00:00:00.123 ERROR viceroy_lib::execute: WebAssembly trapped: Exited with i32 exit status 0
wasm backtrace:
    0: 0x280ea - <unknown>!__wasi_proc_exit
    1: 0x280f6 - <unknown>!_Exit
```

this is unfortunately difficult to test, because the
`Result<(), ExecutionError>` of `ExecuteCtx::run_guest` is spawned in a
task that we do not join on later. (note: that is intentional)

the alternative of a test that checks for a particular log message felt
brittle, and this is a relatively straightfoward change.