Update cache action and unify steps
While looking at the cache issue from #134, I noticed errors in the
GHA output when creating the cache tarball. Furthermore, a newer
version of the cache action existed with some improvements, for
example the ability to use multiple paths in a single step.

This commit bumps the cache action to v2, merging all the steps
together as with the official example, hopefully fixing the errors
and speeding things up.