Add a test fixture for pushing various memory limits.
This fixture, written in WAT to ensure that none of our compilers do
anything too smart, allows the caller to ask it to allocate memory in
various ways: on the wasm heap, in headers, or in the response body. It
also allows the user to ask it to stall for a certain amount of time.
This is handy for testing some edge cases in resource limits.