❓ error: `DictionaryConfigError` variants do not include name
this is a change to the payload(s) of our dictionary config error
variants, specifically removing dictionary names from these variants.

this is a precursor to sharing logic to read the contents of a JSON
dictionary both at startup, and later when a dictionary lookup is
performed.

at startup, these errors are propagated up into a variant that already
includes the name:

```rust
    #[error("invalid configuration for '{name}': {err}")]
    InvalidDictionaryDefinition {
        name: String,
        #[source]
        err: DictionaryConfigError,
    },
```

later on, we can similarly deduce the name of the dictionary given a
handle.