Extend backend TLS configuration with cert host and SNI
To get closer to parity with C@E backend configuration, we add the ability to specify a certificate hostname, distinct from both the host URI or any override host. In addition, backends can now specify whether to use SNI or not; if they do, and the cert hostname is present, it will be used for SNI.

That is slightly less flexible than C@E, which allows distinct SNI and cert hosts. That's not currently possible with the TLS implementation we're using in Viceroy, and the C@E customization is mostly used to _disable_ SNI, which we provide here using a dedicated flag.

The new backend settings are:

* `cert_host` to specify the hostname for certificate checking (and SNI, if enabled)
* `use_sni`, a boolean specifying whether to employ SNI for the backend.

Both settings are optional. If `cert_host` is not specified, the host from the backend URI is used instead. If `use_sni` is not specified, SNI is enabled by default.