Add metadata to viceroy and viceroy-lib Cargo.toml
This adds some extra metadata for both the CLI tool and the library so
that publishes to crates.io won't produce warnings and to make Viceroy
more discoverable. It also points to the correct documentation for the
CLI tool on Fastly's website for the CLI which is how most will interact
with it, whereas the lib points to `docs.rs` so that they know how to
use the library itself.

Closes #17