Bump wasmtime from 0.34.0 to 0.40.1
This updates wasmtime to stay up to date with Compute@Edge in terms of
what version of wasmtime is used. The important change here is that
`ModuleLimits` was removed in version 0.35.0. Lots of the limits we
set before now did not matter and so instead we only needed to set a
few of them with instance limits.