Add CPU time to output for requests
Fastly uses CPU time to denote how much time is spent *actually running*
the WASM code given to us. Given how asynchronous code works, it means
we only run the code when it's being polled. This means we can't just
slap `Instant::now()` in front of the call to the main function of the
WASM in Viceroy. We instead construct a future that will check how long
each call to `poll` was and add it to a `Duration`. When we are done
polling we will have the result of the request made to Viceroy and how
long that took! We then report it with the other metrics for customers
to look at so that they can approximate how much time they spend in
their module. While this is by no means a 1 to 1 in terms of how much
CPU time is actually used as a customer's local environment is not the
same as our POPs it should help guide them in terms of optimizing their
code so that requests are fast.

Sample output when running now for stats:

```
Sep 13 13:22:40.291  INFO request{id=14}: request completed using 1.1 MB of WebAssembly heap
Sep 13 13:22:40.291  INFO request{id=14}: request completed in 5ms
Sep 13 13:22:40.291  INFO request{id=14}: CPU Time 3ms
```

You can see that it takes 5ms for the whole request to complete, but only
3ms of which is spent working on the actual request using the customer's
code.