Implement Async IO Select Hostcalls in Viceroy
This commit adds the fastly_async_io hostcalls `select` and `is_ready`
to Viceroy. Up to this point we only would do `select` for things like
`PendingRequest`. Now we abstract over an `AsyncItem` which could be a
`PendingRequest`, `StreamingBody`, or `Body` and you can call `select`
over any of these async io operations. Most of this code is to
accomodate these changes while maintaining the behavior of `select` for
`PendingRequest` utilizing these new calls and creating the `AsyncItem`
enum to hold all of these differing types in the same `async_item` map.

We also add an integration test similar to the C@E platform's to make
sure that we implement the behavior correctly. With this we can be
confident that select works the way we intend it too! As a bonus side
effect of these changes we also add async support for hosts in our
testing code in Viceroy which gives us some more flexibility in what we
can accomplish or use!