⏩ none should not be defined in cache_override_tag witx
The ABI expects `$none` to have a value of 0, and `$pass` to have a
value of 1.  However, flags in witx start with a value of 1.  Putting
`$none` in the list erroneously defines it to have a value of 1.

Fastly's SDKs do things correctly, but tools like https://github.com/jedisct1/witx-codegen
which generate bindings from witx will generate code that is off by one
bit position.