switch default log level to "error", add -v to run
The previous default log level was "off", which meant that errors
encountered during startup with `viceroy run` were not printed and the
process silently failed.  For example, this might happen if you had an
invalid path in your `fastly.toml` file.

In addition, there was no way to turn on verbose log output with
`viceroy run`, only in serve mode.

This commit moves the default log level to "error", and moves the `-v`
flag into the common set.

Making "error" the default meant that processes that called `proc_exit`
would print WASM stacktraces as errors on exit, but this isn't really an
error from the user's perspective, so we check for that (`I32Exit`) and
suppress the error message in that case only.