Return a ValueAbsent for all the downstream-tls related functions
Currently these functions return a NotAvailable error, I believe this is because we don't support https within Viceroy.

We do support http though, and http is also supported on Fastly, this patch adds an implementation for the downstream-tls functions which is valid for http (non-tls) connections. As Viceroy does not yet support https, we could say this is a full implementation as-of-today. In the future if/when Viceroy supports https, we would need to come back to these functions and make them return the correct, expected values based on the incoming connection (tls or not).