switch most uses of Error::Unsupported to Error::NotAvailable
`Unsupported` is meant to express things that are not supported by the
platform, for example trying to make a request with "HTTP/4".

`NotAvailable` is meant to express things that are not available or
unimplemented in Viceroy.

Previously, `Unsupported` would return `FastlyStatus::Unsupported` and
`NotAvailable` would return the generic `FastlyStatus::Error`.  This
led to a bad user experience for the `NotAvailable` case, as the user
would see an opaque and generic "Error" message rather than
"Unsupported".  The previous commit fixes this by having both
`Unsupported` and `NotAvailable` return `FastlyStatus::Unsupported`.

This commit makes our use of `NotAvailable` consistent.