Upgrade to wasmtime-20, and switch to wasmtime-wasi (#366)
Update the wasmtime dependency to wasmtime-20, and also remove the dependency on
wasi-common in favor of using the preview1-on-preview2 implementation provided
by wasmtime-wasi.

There was also a change in wiggle's borrow checker for the 20.0.0 release that
affected our host call implementations directly: borrows are no longer to a
region, but the whole guest memory. This caused a bit of churn in the wiggle_abi
tree, and I'm not completely certain that all issues have been resolved, as the
test suite can only test so many paths through the host calls. For reference,
here's the wiggle refactoring PR in wasmtime: bytecodealliance/wasmtime#8277.