Allow capturing logging endpoint messages
Some integration tests need to capture log output in order to verify
that logging endpoints work correctly from guests, but it's also
potentially useful for people using Viceroy as a library.

The previous way that the integration tests did this was not reliable
when multiple tests ran in parallel, because there was only one global
hook for capturing logs. Exposing this as a configuration option on
execution contexts instead allows each test to set independent capture
buffers.

I guess this should be considered a breaking API change since
`viceroy_lib::logging::LOG_WRITER` was exported publicly from the crate.