Rename the kv-store interface to object-store in compute.wit (#415)
Rename the kv-store interface in compute.wit to object-store, and add todo!() implementations for the new kv-store interface. The host implementations of those functions will be added in a follow-up PR along with the wiggle implementation for the new host calls, however the adapter has been updated to handle the new hostcalls already.