CI fix: allow build of viceroy-component-adapter
The `generate!` directive `std_feature` generates a `cfg(feature="std")`
guard, but the `viceroy-component-adapter` crate did not have a `std`
feature listed.

The crate does not build with `no_std`; there's a not small amount of
`Vec` littered about. Instead of attempting to make it `no_std` safe,
drop the `std_feature` guard.