ENABLE_EXPERIMENTAL_CACHE_API environment flag
Some of the core cache API hostcalls are incomplete; namely, `insert`
and `lookup` don't respect the "option" arguments. We don't want to
enable access to those APIs in a way that will misbehave (e.g. ignoring
arguments).

Guard the use (and testing) of the incomplete APIs behind an environment
variable, ENABLE_EXPERIMENTAL_CACHE_API. When present, we enable the API
and run the tests.