Guest Profile sample period configuration support
This change matches the configuration support for guest profling
provided by wasmtime and is useful for integrators making use of guest
profiling to be able to get a more fine/coarse grained view of the
execution of their services (and also balance data generated, overhead,
etc.).

In preparing to submit this change for review, I first saw that there
is an alternative implementation of this with
https://github.com/fastly/Viceroy/pull/437, doing mostly the same things
in a slightly different form.