Add CollectingBody for cache use cases (#452)
* Add CollectingBody for cache use cases

`CollectingBody` reads from a `Body` and:

- Stores the data and trailers written, for future delivery in cache
  lookups
- Offers a `read() -> Body` endpoint. The read can proceed concurrently
  with writes, or after the write has completed.

* Address more review comments

* Fix implementation comment

* Fix to revised impl

Missed the "skip already read chunks" step.

* Address some review comments

* Fix broken compile from passing slice

* Use oneshot instead of yield

* Address comments from @drbrain

* Go back to a manual Default impl for CollectingBodyInner

`#[default]` can only be used on unit enums (even if all the enum fields
implement `Default`).

* Test & fix possible deadlock in streaming body

@drbrain pointed out that

> At creation, \[of a tokio `Receiver`\] the initial value is considered seen.

We have to be sure to process that initial value. Switch the read loop
to a do-while form: always process the first value received, then wait
for a future value & loop again.

The test in this commit didn't trigger the problem, but that may be luck
in the scheduler.