Instrument hostcalls when Guest Profiling (#466)
This change matches the behavior of the wasmtime cli by
tying in call-hook instrumentation to record additional
profile data about the stack frames for each hostcall
entry/exit.

This can be very useful in understanding the time guest
code is spent blocked on hostcalls (which won't be sampled
by epoch interrupts).