Ensure `transaction_lookup_async` polls once before returning
Before returning from `transaction_lookup_async`, we attempt to retrieve
data + an obligation from the `CacheKeyObject`, but without blocking for
any outstanding obligations. Only if this fails do we generate a
background task to wait for other obligations to complete.

The net result is that `BusyHandle::pending` will in some cases
immediately be `true` upon return of the `BusyHandle`, e.g. when the
cache needs filling.