🚫 Explicitly reject SERVICE_ID overrides
SERVICE_ID overrides (`on_behalf_of`) are a privileged feature (i.e.
internal to Fastly). Viceroy doesn't implement the relevant behaviors-
and this is an observable difference from normal operation, so we want
to give a clear error if it's requested.

- Add the component-side ABI for service overrides to Viceroy
- Reject requests that come in with a service_id override

If you're internal to Fastly and need this feature in Viceroy, reach out
to me. :)