🙉 'support' (ignore) the sensitive data bit
It turns out there's no API with which to read this bit. It doesn't have
any effect in Viceroy (which doesn't persist the cache anyway), so it
can be safely ignored.

Leave an explicit "we're reading and ignoring this" comment in the
source, to make it clear to future maintainers that it is intentionally
ignored, not a missing feature.