Proptest and directed tests for chunking
A more general set of tests for #479. Assume that the origin can send
chunks of arbitrary length - in the case of #479, a gzip header in a
separate chunk from the content.

These show up as separate chunks in a `hyper::Body`. Ensure that, when
splitting a body at arbitrary chunk lengths, the output still reproduces
the body. This causes failures without #480 (or similar changes).

This validates the fix in #480 (the gzip test fails without #480).