Update Viceroy to the latest WIT files.
This updates Viceroy to the WIT files and component adapter as of
Executed 4319d1e, and updates the host implementations to match the
WIT interfaces.

This also makes the adapter a separate workspace, and updates the build
rules accordingly, to more closely match how this work in the ExecuteD
repository.