Implement and use AsyncRead analogue, for #479
Issue 479 has more details on the "zero-length chunk" issue. It comes
from an API divergence between `http::Body`, which has an explicit "end
of stream" signal, and `std::io::Read` (and its async analogues), which
use "zero bytes read" to indicate an end-of-stream. Providing a
zero-length chunk results in confusion when moving from the HTTP API
to the Read API, as we do in `fastly_http_body::read`.

Modeled after the code in Compute Platform, here we introduce an
analogue to the AsyncRead traits. The helper `check_body_read`
shows how to safely use this to read the whole body, as with AsyncRead
users. The HTTP body read hostcalls are updated to use the new reader,
hiding details of `Body` and sending them through the happy path.

This is an alternative fix for #479, that should cover any lingering
bugs in existing/future `Body` implementations as well.