Remove env var guard for cache features (#488)
Prior to this change, the entry points to the core cache API (`insert`) required that the `ENABLE_EXPERIMENTAL_CACHE_API` be set. This was because not all options were handled; we wouldn't want to complete the hostcall while ignoring an option, without the caller opting in.

However, this means that a user can't use *any* of the core cache API without opting in to the "partially correct" behavior. e.g., a user can't stick to the implemented subset without setting the environment flag.

In this PR, instead, return `NotSupported` for any hostcall options that are not yet implemented. Consume the `options_mask` bits as each option is consumed, and check that the remaining mask is empty at the end of processing.

Note that this is future-compatible when new options are added. If a guest program attempts to invoke an option that is not handled in Viceroy, they'll (correctly) get `NotSupported` with no special changes to Viceroy. Put differently, going forward, Viceroy won't silently discard unknown options bits.

This removes the `ENABLE_EXPERIMENTAL_CACHE_API` environment variable as a guard. A couple of options are still not handled (surrogate keys + range requests), but per the above, they'll return `NotSupported` if invoked.