Refactor get_body to be async
In the most recent builds of the compute platform, get_body can block
until the start of the requested range is available (if the
always_use_requested_range flag is set).

Prepare for this by making get_body async, using a named `impl Future`
that can hold the range-related options.