Avoid "unreachable expression' from proptest
Proptest wraps the body in this syntax:

```rust
let () : () = { ...body here... };
Ok(())
```

This makes the control-flow in the body unusual. The body must return a
unit `()` in the _default_ case, and that default case must be
reachable; but early returns may return `Ok(())` or `Err(...)`.

I don't like this, and should probably complain about it to proptest for
their next major release. But in this commit- just fix it.