Range requests for cache API (#473)
* Refactor get_body to be async

In the most recent builds of the compute platform, get_body can block
until the start of the requested range is available (if the
always_use_requested_range flag is set).

Prepare for this by making get_body async, using a named `impl Future`
that can hold the range-related options.

* Range-request support in CollectingBody

Allow `read_range` in CollectingBody with a start (and optional end)
offset. Add helpers to allow waiting for a known or minimum length,
which we need to get a consistent blocking behavior with Compute.

Add the proptest regressions files for body and collecting_body.

* Range request in CacheEntry

* ABI updates for range requests

* Glue for range request ABI

* Add test for range requests