Move the State to the stack frame of the adapter entrypoint `serve()`.
This is toward getting State out of the heap, where TinyGo (and likely Big Go) clobber it fairly promptly on startup, generally before the first request.

Keep the versions of `get_state_ptr` and `set_state_ptr` that put the ptr in a wasm global. Otherwise, we get a data section in the adapter, which is forbidden.

We will also need to stop allocating our unmanaged stack on the heap and probably just use the existing `$__stack_pointer`.