Remove the custom verify-publishable script.
This script invokes cargo vendor and cargo package in ways that attempt
to simulate how crates are packages for crates.io, however it doesn't
work with the new directory layout.

So instead, replace the script with a simple `cargo publish --dry-run`
invocation. That only checks viceroy-lib and not the viceroy CLI crate,
but it's the bulk of the benefit with a fraction of the complexity.