Core cache: lock FOUND == USABLE (#552)
* Add test of simple cache

* Test expiry in core cache too

* Cache: only return FOUND if USABLE

The Compute platform, at the moment, only returns FOUND if within the
USABLE period (in HTTP semantics, "fresh" or "stale-while-revalidate").
Unfortunately, SDKs rely on this property, and check FOUND or USABLE
variously; sometimes (e.g. in Rust), this results in a
`FOUND`-but-not-`USABLE` result being returned.

We can fix SDKs going forward, but not existing usage. Match Compute's
behavior by locking `FOUND` to `USABLE`. If we want to return expired
results, we'll need more / other bits.